pyproject.toml
[project]
name = "rover"
version = "0.9.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alabaster==0.7.13",
    "altgraph==0.17.3",
    "annotated-types==0.6.0",
    "argparse-addons==0.12.0",
    "babel==2.12.1",
    "bitstruct==8.19.0",
    "breathe==4.35.0",
    "canlib==1.27.606",
    "cantools==39.4.11",
    "certifi==2023.7.22",
    "charset-normalizer==3.2.0",
    "click==8.1.6",
    "common-interfaces-intellisense>=0.1.0",
    "contourpy==1.1.0",
    "crccheck==1.3.0",
    "cycler==0.11.0",
    "diskcache==5.6.3",
    "docutils==0.18.1",
    "fonttools==4.43.0",
    "idna==3.4",
    "imagesize==1.4.1",
    "jinja2==3.1.3",
    "keyboard==0.13.5",
    "kiwisolver==1.4.4",
    "markupsafe==2.1.3",
    "matplotlib==3.9.2",
    "meson>=1.5.2",
    "msgpack==1.0.8",
    "mypy-extensions==1.0.0",
    "ninja>=1.11",
    "nodeenv==1.9.1",
    "numpy==1.26.4",
    "packaging==23.1",
    "pathspec==0.11.2",
    "pillow==10.2.0",
    "platformdirs==3.10.0",
    "pydantic==2.0.2",
    "pydantic-core==2.1.2",
    "pygments==2.15.1",
    "pyinstaller==5.13.2",
    "pyinstaller-hooks-contrib==2023.8",
    "pyparsing==3.0.9",
    "pyright==1.1.382.post1",
    "python-can==4.4.2",
    "python-dateutil==2.8.2",
    "rclpy-intellisense>=1.2.0",
    "requests==2.31.0",
    "rover-py",
    "ruff>=0.9.4",
    "scipy==1.12.0",
    "setuptools==75.1.0",
    "shellcheck-py==0.10.0.1",
    "shfmt-py==3.7.0.1",
    "six==1.16.0",
    "snowballstemmer==2.2.0",
    "sphinx==6.2.1",
    "sphinx-rtd-theme==1.2.2",
    "sphinxcontrib-applehelp==1.0.4",
    "sphinxcontrib-devhelp==1.0.2",
    "sphinxcontrib-htmlhelp==2.0.1",
    "sphinxcontrib-jquery==4.1",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==1.0.3",
    "sphinxcontrib-serializinghtml==1.1.5",
    "textparser==0.24.0",
    "tomli==2.0.1",
    "typing-extensions==4.7.1",
    "urllib3==2.0.7",
    "wrapt==1.16.0",
]

[tool.uv.sources]
rover-py = { path = "rover_py", editable = true }

[tool.ruff.lint]
# see https://docs.astral.sh/ruff/rules
select = [
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-use-pathlib
    "PTH",
    # isort
    "I",
    # pep8-naming
    "N",
]
ignore = ["E501"]
